fix(relay): make reliability counters pass security checks
The first reliability batch passed make verify, which does not run gosec. Pre-push then rejected three G115 findings: a breaker threshold conversion inside a closure and two unchecked narrowing conversions of retry audit numbers.

Convert the breaker threshold immediately after validating its 1..1000 range and use the existing saturating int32 conversion for stream and nonstream attempt metadata. This preserves ordinary request behavior and prevents negative audit numbers after narrowing.

Validation: full pre-push scan with the pinned gosec v2.28.0 built using Go 1.27.1; go test ./platform/grpc ./internal/server ./pkg/safecast; Markdown links; git diff --check. The local scanner was also updated from v2.26.1 to resolve Go 1.27 parsing errors. No new rule suppressions were added.

+++ platform/grpc/service_config.go
@@ -21,14 +21,14 @@ func ServiceBreakerConfig(service string, timeout time.Duration) (*BreakerConfig
 	prefix := "GRPC_" + strings.ToUpper(service) + "_"
 	cfg := DefaultBreakerConfig(service)
 	cfg.FallbackStrategy = FallbackReject
-	samples := 5
+	samples := uint32(5)
 	ratio := 0.6
 	if raw := os.Getenv(prefix + "BREAKER_MIN_REQUESTS"); raw != "" {
 		value, err := strconv.Atoi(raw)
 		if err != nil || value < 1 || value > 1000 {
 			return nil, 0, fmt.Errorf("%sBREAKER_MIN_REQUESTS must be 1..1000", prefix)
 		}
-		samples = value
+		samples = uint32(value)
 	}
 	if raw := os.Getenv(prefix + "BREAKER_FAILURE_RATIO"); raw != "" {
 		value, err := strconv.ParseFloat(raw, 64)
@@ -51,7 +51,7 @@ func ServiceBreakerConfig(service string, timeout time.Duration) (*BreakerConfig
 		}
 	}
 	cfg.ReadyToTrip = func(counts gobreaker.Counts) bool {
-		return counts.Requests >= uint32(samples) && float64(counts.TotalFailures)/float64(counts.Requests) >= ratio
+		return counts.Requests >= samples && float64(counts.TotalFailures)/float64(counts.Requests) >= ratio
 	}
 	return cfg, timeout, nil
 }
